+++ src/composables/PresellService.ts
@@ -28,4 +28,15 @@ export async function getPresell (payload: any) {
     })
 }
 
+export async function deletePresell (payload: any) {
+  return await apiAuth
+    .delete(`affiliate/pre-sell/delete/${payload}`)
+    .then((response: any) => {
+      return response.data.message
+    })
+    .catch((error: any) => {
+      throw error?.response?.data?.message || error?.message
+    })
+}
+
 export default { newPresell, getPresell }