+++ src/models/PresellModel.ts
@@ -1,14 +1,33 @@
 export interface PresellNew {
+  // Template 1 - available fields
   productName: string
-  pageTitleText: string
+  imageId: string
+  affiliateUrl: string
+
+  backColor1: string
+  backColor2: string
+  buttonColor: string
+
+  // Template 2 - available fields
   headLineText: string
   buttonText: string
-  backGradientColor1: string
-  backGradientColor2: string
-  buttonColor: string
-  showLastOffer: boolean
-  showGradientBackground: boolean
-  showAnimationOverButton: boolean
+  buttonChoice: 'image' | 'text'
+  showButtonAnimation: boolean
+
+  // Template 3 - available fields
+  pageTitleText: string
+  serverToHostPresell: string
+  showDetailedInfo: boolean
+  aboutProduct: string
+  howDoesProductWork: string
+  productIngredients: string
+  showLastChanceToBuy: boolean
+  showPromotionValues: boolean
+  promotionValueBefore: number
+  promotionValueWithDiscount: number
+  discountValue: number
+  // Other definitions
+  template: string[]
 }
 
 export enum PresellStatus {